Add assertions to Storage and UI classes
To comply with the A-Assertions level for Week 5

Add an assertion in Storage constructor to ensure filePath parameter
is not empty
Add an assertion in Ui getResponse method to ensure response buffer
contains content.

These cases are should not happen at all and thus should cause
assertion failures and take drastic action as the programme
cannot continue with these conditions. File cannot be read/write
if filePath is empty and bot response cannot be blank.